[NFC][cifuzz] Move Workspace from docker into config_utils.
It doesn't really fit into docker.
