+++ infra/cifuzz/build_fuzzers_entrypoint.py
@@ -17,7 +17,6 @@ import sys
 
 import build_fuzzers
 import config_utils
-import docker
 
 # pylint: disable=c-extension-no-member
 # pylint gets confused because of the relative import of cifuzz.
@@ -54,7 +53,7 @@ def build_fuzzers_entrypoint():
     returncode = 0
   # yapf: disable
   elif build_fuzzers.check_fuzzer_build(
-      docker.Workspace(config),
+      config_utils.Workspace(config),
       config.sanitizer,
       config.language,
       allowed_broken_targets_percentage=config.allowed_broken_targets_percentage
