+++ infra/cifuzz/docker.py
@@ -94,52 +94,3 @@ def _get_args_mapping_host_path_to_container(host_path, container_path=None):
   # --volumes-from) is used for mapping volumes. It will break production.
   container_path = host_path if container_path is None else container_path
   return ['-v', f'{host_path}:{container_path}']
-
-
-class Workspace:
-  """Class representing the workspace directory."""
-
-  def __init__(self, config):
-    self.workspace = config.workspace
-
-  def initialize_dir(self, directory):  # pylint: disable=no-self-use
-    """Creates directory if it doesn't already exist, otherwise does nothing."""
-    os.makedirs(directory, exist_ok=True)
-
-  @property
-  def out(self):
-    """The out directory used for storing the fuzzer build built by
-    build_fuzzers."""
-    # Don't use 'out' because it needs to be used by artifacts.
-    return os.path.join(self.workspace, 'build-out')
-
-  @property
-  def work(self):
-    """The directory used as the work directory for the fuzzer build/run."""
-    return os.path.join(self.workspace, 'work')
-
-  @property
-  def artifacts(self):
-    """The directory used to store artifacts for download by CI-system users."""
-    # This is hardcoded by a lot of clients, so we need to use this.
-    return os.path.join(self.workspace, 'out', 'artifacts')
-
-  @property
-  def clusterfuzz_build(self):
-    """The directory where builds from ClusterFuzz are stored."""
-    return os.path.join(self.workspace, 'cifuzz-prev-build')
-
-  @property
-  def clusterfuzz_coverage(self):
-    """The directory where builds from ClusterFuzz are stored."""
-    return os.path.join(self.workspace, 'cifuzz-prev-coverage')
-
-  @property
-  def coverage_report(self):
-    """The directory where coverage reports generated by cifuzz are put."""
-    return os.path.join(self.workspace, 'cifuzz-coverage')
-
-  @property
-  def corpora(self):
-    """The directory where corpora from ClusterFuzz are stored."""
-    return os.path.join(self.workspace, 'cifuzz-corpus')
