+++ infra/cifuzz/test_helpers.py
@@ -20,7 +20,6 @@ import tempfile
 from unittest import mock
 
 import config_utils
-import docker
 
 
 def _create_config(config_cls, **kwargs):
@@ -53,7 +52,7 @@ def create_workspace(workspace_path='/workspace'):
   """Returns a workspace located at |workspace_path| ('/workspace' by
   default)."""
   config = create_run_config(workspace=workspace_path)
-  return docker.Workspace(config)
+  return config_utils.Workspace(config)
 
 
 def patch_environ(testcase_obj, env=None):
